+++ rbx_reflection_database/src/lib.rs
@@ -45,11 +45,11 @@ mod test {
     }
 
     #[test]
-    fn class_is_a_test() {
+    fn has_superclass_test() {
         let database = get();
         let part_class_descriptor = database.classes.get("Part").unwrap();
         let instance_class_descriptor = database.classes.get("Instance").unwrap();
-        assert!(database.class_is_a(part_class_descriptor, instance_class_descriptor));
-        assert!(!database.class_is_a(instance_class_descriptor, part_class_descriptor));
+        assert!(database.has_superclass(part_class_descriptor, instance_class_descriptor));
+        assert!(!database.has_superclass(instance_class_descriptor, part_class_descriptor));
     }
 }